separated edit dialog from medical bio; started to refactor Medical Routes and views
