+++ src/components/routes/MedicalRoutes.js
@@ -5,7 +5,7 @@ import PetRepository from "../../repositories/PetRepository";
 import { useHistory } from "react-router-dom";
 import { Route } from "react-router-dom";
 import { ViewMedical } from "../medicals/ViewMedical";
-import { PetMedicalBio } from "../medicals/PetMedicalBio"
+import { PetMedicalBio } from "../medicals/bio/PetMedicalBio"
 import { IncidentsList } from "../medicals/IncidentsList"
 import { Incident } from "../medicals/Incident";
 import { VetVisitsList } from "../medicals/VetVisitsList";
@@ -13,55 +13,7 @@ import { VetVisit } from "../medicals/VetVisit";
 import { VaccinationsList } from "../medicals/VaccinationsList";
 import { MedicationsList } from "../medicals/MedicationsList";
 
-export const MedicalRoutes = () => {
-    const [user, updateUser] = useState({});
-    const [pet, setPet] = useState({});
-    const [myPets, setMyPets] = useState([]);
-
-    const { getCurrentUser } = useSimpleAuth();
-    const history = useHistory()
-    
-    const syncPet = () => {
-        if(user.defaultPetId) {
-            PetRepository.getExpandAll(user.defaultPetId).then(
-                (data) => {
-                    setPet(data);
-                }
-            );
-        }
-    };
-
-    const syncPets = () => {
-        PetRepository.getAllExpandAllByUser(getCurrentUser().id).then(
-            (data) => {
-                data.sort((el1) => {
-                    if (el1.id === user.defaultPetId) {
-                        return -1;
-                    }
-                });
-                setMyPets(data);
-            }
-        );
-    };
-
-    const syncUser = () => {
-        UserRepository.get(getCurrentUser().id).then((data) => {
-            updateUser(data)
-        });
-    };
-
-    useEffect(() => {
-        syncPet()
-        syncPets()
-    },[user])
-
-    useEffect(() => {
-        syncUser();
-    }, []);
-
-    useEffect(() => {
-        syncUser();
-    }, [history.location.state]);
+export const MedicalRoutes = ({pet, syncPets}) => {
 
     return (
         <>
